+++ src/Fauna/Output/herbivore_data.cpp
@@ -159,4 +159,3 @@ HerbivoreData HerbivoreData::create_datapoint(
 
   return result;
 }
-