+++ src/Fauna/get_forage_demands.h
@@ -13,8 +13,22 @@
 
 namespace Fauna {
 // Forward Declarations
+class GivenPointAllometry;
 class Hft;
 enum class Sex;
+
+/// Extrapolate/Interpolate allometrically from a value for adult male.
+/**
+ * \param allometry Parameters to define the allometric relationship.
+ * \param bodymass_male_adult The body mass of an adult, male animal. See
+ * \ref Hft::body_mass_male.
+ * \param bodymass Current body mass of the animal for which the value is to be
+ * calculated.
+ * \throw std::invalid_argument If any given parameter is outside of the
+ * allowed range.
+ */
+double calc_allometry(const GivenPointAllometry& allometry,
+                      const double bodymass_male_adult, const double bodymass);
 
 /// Function object to calculate forage demands for a herbivore.
 /**
@@ -25,7 +39,7 @@ class GetForageDemands {
   /// Constructor.
   /**
    * \param hft Herbivore functional type.
-   * \param sex The sex of the herbivore individual/cohort.
+   * \param sex The sex of the herbivore cohort.
    * \throw std::invalid_argument If `hft==NULL`.
    */
   GetForageDemands(std::shared_ptr<const Hft> hft, const Sex sex);
@@ -182,7 +196,7 @@ class GetForageDemands {
   double energy_needs;                 /// [MJ/ind]
   ForageMass max_intake;               /// [kgDM/ind/day]
   int today;                           /// 0 = January 1st
-  /** @} */
+                                       /** @} */
 };
 
 }  // namespace Fauna