+++ src/Fauna/hft.cpp
@@ -28,8 +28,7 @@ bool Hft::is_valid(const Parameters& params, std::string& msg) const {
   }
 
   //------------------------------------------------------------
-  if (params.herbivore_type == HerbivoreType::Cohort ||
-      params.herbivore_type == HerbivoreType::Individual) {
+  if (params.herbivore_type == HerbivoreType::Cohort) {
     if (body_fat_birth <= 0.0) {
       stream << "body_fat.birth must be >0.0 (" << body_fat_birth << ")"
              << std::endl;
@@ -185,16 +184,6 @@ bool Hft::is_valid(const Parameters& params, std::string& msg) const {
       is_valid = false;
     }
 
-    if (params.herbivore_type == HerbivoreType::Individual &&
-        establishment_density <= 2.0 / params.habitat_area_km2) {
-      stream << "establishment.density (" << establishment_density
-             << " ind/km²) "
-             << "must not be smaller than two individuals in a habitat"
-             << " (habitat_area_km2 = " << params.habitat_area_km2 << " km²)."
-             << std::endl;
-      is_valid = false;
-    }
-
     if (expenditure_components.empty()) {
       stream << "No energy expenditure components defined." << std::endl;
       is_valid = false;
@@ -217,12 +206,38 @@ bool Hft::is_valid(const Parameters& params, std::string& msg) const {
       is_valid = false;
     }
 
-    if (expenditure_components.count(ExpenditureComponent::Allometric) &&
-        expenditure_allometric.coefficient < 0.0) {
-      stream << "Coefficient for allometric expenditure must not be "
-                "negative. That would result in negative expenditure values. "
-                "Current value: expenditure_allometric_coefficient = "
-             << expenditure_allometric.coefficient << std::endl;
+    if (expenditure_components.count(
+            ExpenditureComponent::BasalMetabolicRate) &&
+        expenditure_components.count(
+            ExpenditureComponent::FieldMetabolicRate)) {
+      stream << "Both \"BasalMetabolicRate\" and \"FieldMetabolicRate\" are "
+             << "chosen as expenditure components, but \"FieldMetabolicRate\" "
+             << "already includes the basal metabolic rate." << std::endl;
+      is_valid = false;
+    }
+
+    if (expenditure_components.count(
+            ExpenditureComponent::BasalMetabolicRate) ||
+        expenditure_components.count(
+            ExpenditureComponent::FieldMetabolicRate)) {
+      if (expenditure_basal_rate.value_male_adult <= 0.0) {
+        stream << "Daily expenditure (MJ/day) in an adult male in "
+               << "`expenditure.basal_rate.value_male_adult` must be a "
+               << "positive number"
+               << " (current value: " << expenditure_basal_rate.value_male_adult
+               << ")" << std::endl;
+        is_valid = false;
+      }
+    }
+
+    if (expenditure_components.count(
+            ExpenditureComponent::FieldMetabolicRate) &&
+        expenditure_fmr_multiplier < 1.0) {
+      stream << "The field metabolic rate (FMR) multiplier "
+             << "`expenditure.fmr_multiplier` must be >= 1 because FMR is "
+             << "larger than BMR."
+             << " (current value: " << expenditure_fmr_multiplier << ")"
+             << std::endl;
       is_valid = false;
     }
 
@@ -264,21 +279,23 @@ bool Hft::is_valid(const Parameters& params, std::string& msg) const {
     }
 
     if (digestion_limit == DigestiveLimit::Allometric &&
-        digestion_allometric.coefficient < 0.0) {
-      stream << "Coefficient in `digestion.allometric` must not be negative "
-                "if 'Allometric' is set as a digestive limit."
-             << " (current value: " << digestion_allometric.coefficient << ")"
-             << std::endl;
+        (digestion_allometric.value_male_adult <= 0.0 ||
+         digestion_allometric.value_male_adult >= 1.0)) {
+      stream << "Dry-matter intake as fraction of body mass in an adult male "
+             << "in `digestion.value_male_adult` must be between "
+             << "in the interval (0,1)."
+             << " (current value: " << digestion_allometric.value_male_adult
+             << ")" << std::endl;
       is_valid = false;
     }
 
     if (digestion_limit == DigestiveLimit::FixedFraction &&
         (digestion_fixed_fraction <= 0.0 || digestion_fixed_fraction >= 1.0)) {
-      stream
-          << "Body mass fraction `digestion.fixed_fraction` must be in "
-             "interval (0,1) if 'FixedFraction' is set as the digestive limit."
-          << " (current value: " << digestion_fixed_fraction << ")"
-          << std::endl;
+      stream << "Body mass fraction `digestion.fixed_fraction` must be in "
+                "interval (0,1) if 'FixedFraction' is set as the digestive "
+                "limit."
+             << " (current value: " << digestion_fixed_fraction << ")"
+             << std::endl;
       is_valid = false;
     }
 
@@ -399,15 +416,6 @@ bool Hft::is_valid(const Parameters& params, std::string& msg) const {
       }
     }
     // add more checks in alphabetical order
-  }
-
-  if (params.herbivore_type == HerbivoreType::Individual) {
-    if (mortality_factors.count(MortalityFactor::StarvationIlliusOConnor2000)) {
-      stream << "Mortality factor `StarvationIlliusOConnor2000` "
-                "is not meant for individual mode."
-             << std::endl;
-      is_valid = false;
-    }
   }
 
   // convert stream to string