+++ src/Fauna/insfile_reader.cpp
@@ -59,8 +59,7 @@ InsfileReader::InsfileReader(const std::string filename)
       throw std::runtime_error("Parameters are not valid:\n" + err_msg);
   }
 
-  if (params.herbivore_type == HerbivoreType::Cohort ||
-      params.herbivore_type == HerbivoreType::Individual) {
+  if (params.herbivore_type == HerbivoreType::Cohort) {
     auto hft_table_array = ins->get_table_array("hft");
     if (hft_table_array) {
       for (const auto& hft_table : *hft_table_array) {
@@ -320,8 +319,12 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
     assert(array);
     hft.expenditure_components = {};
     for (const auto& i : *array)
-      if (lowercase(i) == lowercase("Allometric"))
-        hft.expenditure_components.insert(ExpenditureComponent::Allometric);
+      if (lowercase(i) == lowercase("BasalMetabolicRate"))
+        hft.expenditure_components.insert(
+            ExpenditureComponent::BasalMetabolicRate);
+      else if (lowercase(i) == lowercase("FieldMetabolicRate"))
+        hft.expenditure_components.insert(
+            ExpenditureComponent::FieldMetabolicRate);
       else if (lowercase(i) == lowercase("Taylor1981"))
         hft.expenditure_components.insert(ExpenditureComponent::Taylor1981);
       else if (lowercase(i) == lowercase("Thermoregulation"))
@@ -330,9 +333,9 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
       else if (lowercase(i) == lowercase("Zhu2018"))
         hft.expenditure_components.insert(ExpenditureComponent::Zhu2018);
       else
-        throw invalid_option(
-            hft, "expenditure.components", i,
-            {"Allometric", "Taylor1981", "Thermoregulation", "Zhu2018"});
+        throw invalid_option(hft, "expenditure.components", i,
+                             {"BasalMetabolicRate", "FieldMetabolicRate",
+                              "Taylor1981", "Thermoregulation", "Zhu2018"});
     if (table->get_table("expenditure"))
       table->get_table("expenditure")->erase("components");
   }
@@ -487,8 +490,8 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
   {
     const bool mandatory = hft.digestion_limit == DigestiveLimit::Allometric;
     const auto value = find_hft_parameter<double>(
-        table, "digestion.allometric.coefficient", mandatory);
-    if (value) hft.digestion_allometric.coefficient = *value;
+        table, "digestion.allometric.fraction_male_adult", mandatory);
+    if (value) hft.digestion_allometric.value_male_adult = *value;
   }
   {
     const bool mandatory = (hft.digestion_limit == DigestiveLimit::Allometric);
@@ -497,18 +500,23 @@ Hft InsfileReader::read_hft(const std::shared_ptr<cpptoml::table>& table) {
     if (value) hft.digestion_allometric.exponent = *value;
   }
   {
-    const bool mandatory =
-        (hft.expenditure_components.count(ExpenditureComponent::Allometric));
-    const auto value = find_hft_parameter<double>(
-        table, "expenditure.allometric.coefficient", mandatory);
-    if (value) hft.expenditure_allometric.coefficient = *value;
+    const bool mandatory = (hft.expenditure_components.count(
+                                ExpenditureComponent::BasalMetabolicRate) ||
+                            hft.expenditure_components.count(
+                                ExpenditureComponent::FieldMetabolicRate));
+    const auto value1 = find_hft_parameter<double>(
+        table, "expenditure.basal_rate.mj_per_day_male_adult", mandatory);
+    if (value1) hft.expenditure_basal_rate.value_male_adult = *value1;
+    const auto value2 = find_hft_parameter<double>(
+        table, "expenditure.basal_rate.exponent", mandatory);
+    if (value2) hft.expenditure_basal_rate.exponent = *value2;
   }
   {
-    const bool mandatory =
-        (hft.expenditure_components.count(ExpenditureComponent::Allometric));
+    const bool mandatory = hft.expenditure_components.count(
+        ExpenditureComponent::FieldMetabolicRate);
     const auto value = find_hft_parameter<double>(
-        table, "expenditure.allometric.exponent", mandatory);
-    if (value) hft.expenditure_allometric.exponent = *value;
+        table, "expenditure.fmr_multiplier", mandatory);
+    if (value) hft.expenditure_fmr_multiplier = *value;
   }
   {
     const bool mandatory =
@@ -748,14 +756,6 @@ void InsfileReader::read_table_simulation() {
     }
   }
   {
-    const auto key = "simulation.habitat_area_km2";
-    auto value = ins->get_qualified_as<double>(key);
-    if (value) {
-      params.habitat_area_km2 = *value;
-      remove_qualified_key(ins, key);
-    }
-  }
-  {
     const auto key = "simulation.establishment_interval";
     auto value = ins->get_qualified_as<int>(key);
     if (value) {
@@ -769,10 +769,8 @@ void InsfileReader::read_table_simulation() {
     if (value) {
       if (lowercase(*value) == lowercase("Cohort"))
         params.herbivore_type = HerbivoreType::Cohort;
-      else if (lowercase(*value) == lowercase("Individual"))
-        params.herbivore_type = HerbivoreType::Individual;
       else
-        throw invalid_option(key, *value, {"Cohort", "Individual"});
+        throw invalid_option(key, *value, {"Cohort"});
       remove_qualified_key(ins, key);
     } else
       throw missing_parameter(key);